make folder plugin optional
