+++ src/main/java/org/jenkinsci/lib/configprovider/model/Config.java
@@ -23,24 +23,13 @@
  */
 package org.jenkinsci.lib.configprovider.model;
 
-import com.cloudbees.hudson.plugins.folder.AbstractFolder;
 import edu.umd.cs.findbugs.annotations.NonNull;
-import edu.umd.cs.findbugs.annotations.Nullable;
 import hudson.model.*;
-import jenkins.model.Jenkins;
 import org.jenkinsci.lib.configprovider.ConfigProvider;
-import org.jenkinsci.plugins.configfiles.ConfigFileStore;
-import org.jenkinsci.plugins.configfiles.buildwrapper.ManagedFileUtil;
-import org.jenkinsci.plugins.configfiles.folder.FolderConfigFileProperty;
-import org.jenkinsci.plugins.configfiles.GlobalConfigFiles;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
 
 import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.logging.Level;
 import java.util.logging.Logger;
 
 /**
@@ -52,92 +41,7 @@ import java.util.logging.Logger;
  */
 public class Config implements Serializable, Describable<Config> {
 
-    private final static Logger LOGGER = Logger.getLogger(Config.class.getName());
-
-    @NonNull
-    public static List<Config> getConfigsInContext(@Nullable ItemGroup itemGroup, Class<? extends Descriptor> descriptor) {
-
-        List<Config> configs = new ArrayList<Config>();
-
-        while (itemGroup != null) {
-            if (itemGroup instanceof AbstractFolder) {
-
-                final AbstractFolder<?> folder = AbstractFolder.class.cast(itemGroup);
-                ConfigFileStore store = folder.getProperties().get(FolderConfigFileProperty.class);
-                if (store != null) {
-                    if (descriptor == null) {
-                        configs.addAll(store.getConfigs());
-                    } else {
-                        configs.addAll(store.getConfigs(descriptor));
-                    }
-                }
-            }
-            if (itemGroup instanceof Item) {
-                itemGroup = Item.class.cast(itemGroup).getParent();
-            }
-            if (itemGroup instanceof Jenkins) {
-                // we are on top scope...
-                if (descriptor == null) {
-                    configs.addAll(GlobalConfigFiles.get().getConfigs());
-                } else {
-                    configs.addAll(GlobalConfigFiles.get().getConfigs(descriptor));
-                }
-                itemGroup = null;
-            }
-        }
-        return configs;
-    }
-
-    public static <T extends Config> T getByIdOrNull(@Nullable ItemGroup itemGroup, @NonNull String configId) {
-
-        while (itemGroup != null) {
-            if (itemGroup instanceof AbstractFolder) {
-                final AbstractFolder<?> folder = AbstractFolder.class.cast(itemGroup);
-                ConfigFileStore store = folder.getProperties().get(FolderConfigFileProperty.class);
-                if (store != null) {
-                    Config config = store.getById(configId);
-                    if (config != null) {
-                        return (T) config;
-                    }
-                }
-            }
-            if (itemGroup instanceof Item) {
-                itemGroup = Item.class.cast(itemGroup).getParent();
-            }
-            if (itemGroup instanceof Jenkins) {
-                // we are on top scope...
-                return (T) GlobalConfigFiles.get().getById(configId);
-            } else {
-                continue;
-            }
-        }
-
-        return null;
-    }
-
-    public static <T extends Config> T getByIdOrNull(@NonNull Item item, @NonNull String configId) {
-        if (item instanceof AbstractFolder) {
-            // configfiles defined in the folder should be available in the context of the folder
-            return (T) getByIdOrNull((ItemGroup) item, configId);
-        }
-        if (item != null) {
-            LOGGER.log(Level.FINE, "try with: " + item.getParent());
-            return (T) getByIdOrNull(item.getParent(), configId);
-        }
-        return null;
-    }
-
-    public static <T extends Config> T getByIdOrNull(@NonNull Run<?, ?> build, @NonNull String configId) {
-        Item parent = build.getParent();
-        Config configFile;
-        if (parent instanceof ItemGroup) {
-            configFile = Config.getByIdOrNull((ItemGroup) parent, configId);
-        } else {
-            configFile = Config.getByIdOrNull(parent, configId);
-        }
-
-        return (T) configFile;
-    }
+    private static final Logger LOGGER = Logger.getLogger(Config.class.getName());
 
     /**
      * a unique id along all providers!
