+++ src/main/java/org/jenkinsci/plugins/configfiles/maven/job/MvnGlobalSettingsProvider.java
@@ -1,7 +1,6 @@
 package org.jenkinsci.plugins.configfiles.maven.job;
 
 import hudson.Extension;
-import hudson.ExtensionList;
 import hudson.FilePath;
 import hudson.model.Item;
 import hudson.model.ItemGroup;
@@ -9,20 +8,18 @@ import hudson.model.TaskListener;
 import hudson.model.AbstractBuild;
 
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import hudson.util.ListBoxModel;
-import jenkins.model.Jenkins;
 import jenkins.mvn.GlobalSettingsProvider;
 import jenkins.mvn.GlobalSettingsProviderDescriptor;
 
 import org.apache.commons.lang.StringUtils;
 import org.jenkinsci.lib.configprovider.model.Config;
+import org.jenkinsci.plugins.configfiles.ConfigFiles;
 import org.jenkinsci.plugins.configfiles.buildwrapper.ManagedFileUtil;
 import org.jenkinsci.plugins.configfiles.common.CleanTempFilesAction;
 import org.jenkinsci.plugins.configfiles.maven.GlobalMavenSettingsConfig;
@@ -71,11 +68,11 @@ public class MvnGlobalSettingsProvider extends GlobalSettingsProvider {
 
             Config c = null;
             if (build instanceof Item) {
-                c = Config.getByIdOrNull((Item) build, settingsConfigId);
+                c = ConfigFiles.getByIdOrNull((Item) build, settingsConfigId);
             } else if (build instanceof ItemGroup) {
-                c = Config.getByIdOrNull((ItemGroup) build, settingsConfigId);
+                c = ConfigFiles.getByIdOrNull((ItemGroup) build, settingsConfigId);
             } else if (build.getParent() instanceof ItemGroup) {
-                c = Config.getByIdOrNull((ItemGroup) build.getParent(), settingsConfigId);
+                c = ConfigFiles.getByIdOrNull((ItemGroup) build.getParent(), settingsConfigId);
             }
 
 
@@ -138,7 +135,7 @@ public class MvnGlobalSettingsProvider extends GlobalSettingsProvider {
         public ListBoxModel doFillSettingsConfigIdItems(@AncestorInPath ItemGroup context) {
             ListBoxModel items = new ListBoxModel();
             items.add("please select", "");
-            for (Config config : Config.getConfigsInContext(context, GlobalMavenSettingsConfigProvider.class)) {
+            for (Config config : ConfigFiles.getConfigsInContext(context, GlobalMavenSettingsConfigProvider.class)) {
                 items.add(config.name, config.id);
             }
             return items;
